CMS-761: Move PaginationBar to components dir
